Enable Ruby 1.9.3
Relaxes Ruby version requirement to ">= 1.9.3"
(was ">= 2.3.3").

Also edits .travis.yml to test on Ruby 1.9.3

---

[DEMO commit only; removes some Travis tests
of other Ruby versions, since they are irrelevant
to checking support of *older* Ruby.

DO NOT MERGE THIS COMMIT.

This can be fixed up to be proper and mergeable
by re-adding those older Ruby versions to .travis.yml.]