Enable Ruby 2.1.5
Relaxes Ruby version requirement to ">= 2.1.5"
(was ">= 2.3.3").

Also edits .travis.yml to test on Ruby 2.1.5

---

[DEMO commit only; removes some Travis tests
of other Ruby versions, since they are irrelevant
to checking support of *older* Ruby.

DO NOT MERGE THIS COMMIT.

This can be fixed up to be proper and mergeable
by re-adding those olfer Ruby versions to .travis.yml.]